+++ Source/Meadow F7 Feather/Maple/MeadowMapleLed/Controllers/LedController.cs
@@ -96,7 +96,7 @@ namespace MeadowMapleLed.Controller
         protected Color GetRandomColor()
         {
             var random = new Random();
-            return Color.FromHsba(random.NextDouble(), 1, 1);
+            return Color.FromHsba((float)random.NextDouble(), 1, 1);
         }
     }
 }
